+++ example/prismarine/common/myobject/models.py
@@ -5,7 +5,7 @@ from prismarine.runtime import Cluster
 c = Cluster('MyAppWithPrismarine')
 
 
-@c.model(PK='Foo', SK='Bar')
+@c.model(PK='Foo', SK='Bar', trigger='itemlogger')
 class Item(TypedDict):
     Foo: str
     Bar: str